Mostly just fixes.
    - Fix position of score message in follow mode.
    - Fix when one-shot bonus is triggered.
    - Some more timing fixes.
    - Make "Fighting Gorillas" the default track.
    - Bonuses are now float multipliers.

+++ Classes/GorillasConfig.h
@@ -82,8 +82,8 @@
 @property (readwrite, assign) NSDictionary  *topScoreHistory;
 @property (readwrite) int                   missScore;
 @property (readwrite) int                   killScore;
-@property (readwrite) int                   bonusOneShot;
-@property (readwrite) int                   bonusSkill;
+@property (readwrite) float                 bonusOneShot;
+@property (readwrite) float                 bonusSkill;
 @property (readwrite) int                   deathScoreRatio;
 @property (readonly) int                    deathScore;
